requirements.txt
Flask==2.3.3
fireworks-ai==0.1.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Werkzeug==2.3.7


